djangoEcommerce-master/ecom/project/ecommerce/requirements.txt
appdirs==1.4.4
apturl==0.5.2
asgiref==3.4.1
autopep8==1.5.6
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.10.0
blinker==1.4
Brlapi==0.7.0
camelot-py==0.10.1
certifi==2019.11.28
cffi==1.14.6
chardet==4.0.0
cliapp==1.20180812.1
click==8.0.1
cmdtest==0.32+git
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2.5
django-ckeditor==6.2.0
django-cleanup==5.2.0
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-filter==2.4.0
django-js-asset==2.0.0
djangorestframework==3.12.4
docopt==0.6.2
duplicity==0.8.12.0
entrypoints==0.3
et-xmlfile==1.1.0
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
ghostscript==0.7
httplib2==0.14.0
humanfriendly==9.2
idna==2.8
importlib-metadata==4.8.1
Js2Py==0.71
keyring==18.0.1
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.0
matplotlib==3.4.3
monotonic==1.5
MouseInfo==0.1.3
netifaces==0.10.4
numpy==1.21.2
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.1.48
openpyxl==3.0.7
packaging==20.3
pandas==1.3.2
paramiko==2.6.0
pbr==5.8.1
pdf2image==1.16.0
pdfminer.six==20201018
pdftopng==0.2.3
pexpect==4.6.0
phonenumbers==8.12.31
Pillow==8.3.1
pipwin==0.5.1
protobuf==3.6.1
PyAudio==0.2.11
PyAutoGUI==0.9.53
pycairo==1.16.2
pycodestyle==2.7.0
pycparser==2.20
pycups==1.9.73
PyGetWindow==0.0.9
Pygments==2.3.1
PyGObject==3.36.0
pyjokes==0.6.0
pyjsparser==2.7.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyNaCl==1.3.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyperclip==1.8.2
PyPrind==2.11.3
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyreadline==2.1
PyRect==0.1.4
pyRFC3339==1.1
PyScreeze==0.1.28
pySmartDL==1.3.4
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python3-xlib==0.15
pyttsx3==2.90
pytweening==1.0.4
pytz==2021.1
pywhatkit==5.1
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
sqlparse==0.4.1
systemd-python==234
tabulate==0.8.9
testresources==2.0.1
toml==0.10.2
torch==1.8.0
torchvision==0.9.0
tqdm==4.43.0
ttystatus==0.38
typing==3.7.4.3
typing-extensions==3.10.0.0
tzlocal==3.0
ubuntu-advantage-tools==27.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.4.4
wadllib==1.3.3
wikipedia==1.4.0
xkit==0.0.0
xlrd==0.7.1
XlsxWriter==1.2.8
xlwt==1.3.0
zipp==3.5.0
